map crashing

+++ src/parsing/parsing.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gumendes <gumendes@student.42porto.com>    +#+  +:+       +#+        */
+/*   By: david-fe <david-fe@student.42.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/16 11:14:54 by gumendes          #+#    #+#             */
-/*   Updated: 2025/10/09 14:09:22 by gumendes         ###   ########.fr       */
+/*   Updated: 2025/10/09 16:23:18 by david-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ typedef struct s_map
 	unsigned int	floor;
 	unsigned int	ceiling;
 	char			**map;
-	float			playersx;
-	float			playersy;
+	double			playersx;
+	double			playersy;
 	char			playero;
 	int				*max_x;
 } t_map;
